variable de entorno y readme actualizados
