+++ requirements.txt
@@ -4,3 +4,4 @@ sqlalchemy
 asyncpg
 psycopg2-binary
 greenlet
+python-dotenv
